requirements.txt
pandas>=2.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
